requirements.txt
flask
flask-wtf
flask-limiter
sqlalchemy
pymysql
passlib[bcrypt]
python-dotenv
